Proper Uninstall
Refactor Deactivation class to utilize Cleanup class for data cleanup and remove redundant methods. Update constant reference for cleaning key.
